+++ webapp/api/v1/interviews/views/__init__.py
@@ -1,17 +1,14 @@
+from .abort_recording_view import AbortRecordingView
 from .behavior_analysis_list_view import BehaviorAnalysisListView
+from .complete_recording_view import CompleteRecordingView
 from .finish_interview_view import FinishInterviewView
 from .generate_analysis_report_view import GenerateAnalysisReportView
+from .initiate_recording_view import InitiateRecordingView
 from .interview_analysis_report_view import InterviewAnalysisReportView
 from .interview_session_viewset import InterviewSessionViewSet
-from .recording_views import (
-  AbortRecordingView,
-  CompleteRecordingView,
-  InitiateRecordingView,
-  PlaybackUrlView,
-  RecordingListView,
-  UploadPartView,
-  UploadRecordingView,
-)
+from .playback_url_view import PlaybackUrlView
+from .presign_part_view import PresignPartView
+from .recording_list_view import RecordingListView
 from .start_interview_view import StartInterviewView
 from .submit_answer_view import SubmitAnswerView
 from .turn_list_view import InterviewTurnListView
@@ -27,9 +24,8 @@ __all__ = [
   "InterviewSessionViewSet",
   "InterviewTurnListView",
   "PlaybackUrlView",
+  "PresignPartView",
   "RecordingListView",
   "StartInterviewView",
   "SubmitAnswerView",
-  "UploadPartView",
-  "UploadRecordingView",
 ]
