+++ webapp/config/settings/components/common.py
@@ -24,6 +24,10 @@ ASGI_APPLICATION = "config.asgi.application"
 
 DEVELOPER = env.str("DEVELOPER", "Mefit")
 
+# 파일 업로드 제한 (기본 2.5MB → 10MB로 확장)
+FILE_UPLOAD_MAX_MEMORY_SIZE = 10 * 1024 * 1024  # 10MB
+DATA_UPLOAD_MAX_MEMORY_SIZE = 10 * 1024 * 1024  # 10MB
+
 __all__ = [
   "env",
   "BASE_DIR",
@@ -34,4 +38,6 @@ __all__ = [
   "SITE_ID",
   "WSGI_APPLICATION",
   "ASGI_APPLICATION",
+  "FILE_UPLOAD_MAX_MEMORY_SIZE",
+  "DATA_UPLOAD_MAX_MEMORY_SIZE",
 ]
